+++ javaparser-core/src/main/java/com/github/javaparser/ast/expr/MarkerAnnotationExpr.java
@@ -31,7 +31,7 @@ import com.github.javaparser.TokenRange;
 import java.util.function.Consumer;
 import java.util.Optional;
 import com.github.javaparser.ast.Generated;
-import static com.github.javaparser.JavaParser.parseName;
+import static com.github.javaparser.StaticJavaParser.parseName;
 
 /**
  * An annotation that uses only the annotation type name.
@@ -39,7 +39,7 @@ import static com.github.javaparser.JavaParser.parseName;
  *
  * @author Julio Vilmar Gesser
  */
-public final class MarkerAnnotationExpr extends AnnotationExpr {
+public class MarkerAnnotationExpr extends AnnotationExpr {
 
     public MarkerAnnotationExpr() {
         this(null, new Name());